* commit 'f5cde447343445be2d83e54b138df8da94301acf':
  TSUGI-114 Fix bug when no questions

+++ instructor-home.php
@@ -28,10 +28,12 @@ if (!$toolTitle) {
 
 $questions = $QW_DAO->getQuestions($_SESSION["qw_id"]);
 
-// Clear any preview responses
-$instructors = $QW_DAO->findInstructors($CONTEXT->id);
-foreach($instructors as $instructor) {
-    $QW_DAO->deleteAnswers($questions, $instructor["user_id"]);
+// Clear any preview responses if there are questions
+if ($questions) {
+    $instructors = $QW_DAO->findInstructors($CONTEXT->id);
+    foreach($instructors as $instructor) {
+        $QW_DAO->deleteAnswers($questions, $instructor["user_id"]);
+    }
 }
 
 include("menu.php");
